+++ R/get_channel_stats.R
@@ -14,6 +14,9 @@
 #' 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' get_channel_stats(channel_id="UChTJTbr5kf3hYazJZO-euHg")
 #' }