+++ R/get_stats.R
@@ -11,6 +11,9 @@
 #' 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' get_stats(video_id="N708P-A45D0")
 #' }