+++ R/get_subscriptions.R
@@ -21,6 +21,9 @@
 #' 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' get_subscriptions(filter=c(channel_id="UChTJTbr5kf3hYazJZO-euHg"))
 #' }