+++ R/get_video_details.R
@@ -14,6 +14,9 @@
 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' get_video_details(video_id="yJXTXN4xrI8")
 #' }