+++ R/list_caption_tracks.R
@@ -18,6 +18,9 @@
 #' 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' list_caption_tracks(video_id="yJXTXN4xrI8")
 #' }