+++ R/list_channel_activities.R
@@ -21,6 +21,9 @@
 #' 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' list_channel_activities(filter = c(channel_id="UCRw8bIz2wMLmfgAgWm903cA"))
 #' list_channel_activities(filter = c(channel_id="UCRw8bIz2wMLmfgAgWm903cA", regionCode="US"))
 #' }