+++ R/list_channel_resources.R
@@ -22,6 +22,9 @@
 #' @examples
 #' 
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' list_channel_resources(filter = c(channel_id = "UCT5Cx1l4IS3wHkJXNyuj4TA"))
 #' list_channel_resources(filter = c(username = "latenight"), part="id, contentDetails")
 #' list_channel_resources(filter = c(username = "latenight"), part="id, contentDetails",