+++ R/list_channel_videos.R
@@ -11,6 +11,9 @@
 #' @examples
 #' 
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' list_channel_videos(channel_id = "UCT5Cx1l4IS3wHkJXNyuj4TA")
 #' list_channel_videos(channel_id = "UCT5Cx1l4IS3wHkJXNyuj4TA", max_results=10)
 #' }