+++ R/list_langs.R
@@ -1,13 +1,19 @@
 #' List Languages That Youtube Currently Supports
 #' 
-#' @return data.frame with 3 columns: hl (two letter abbreviation), name (of the language), etag
 #' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
+#' @return data.frame with 3 columns: hl (two letter abbreviation), name (of the language), etag
+#' 
 #' @export
+#' 
 #' @references \url{https://developers.google.com/youtube/v3/docs/i18nLanguages/list}
+#' 
 #' @examples
 #' \dontrun{
+#' 
+#' # Set API token via yt_oauth() first
+#' 
 #' list_langs()
 #' }