+++ NEWS.md
@@ -1,5 +1,6 @@
 # version 0.6.0 
 
+* Based on CRAN feedback, add comment about yt_outh to all man pages
 * video_id is returned as part of the list for get_stats, get_video_details
 * handles errors stemming from bad video id for get_stats, get_video_details
 * fixed bug in get_comment that delivers separate results for diff. filters, error handling for bad comment_id, and now comment_id returned as part of df